+++ comics/comics/picturesforsadchildren.py
@@ -1,4 +1,4 @@
-from comics.aggregator.crawler import CrawlerBase, CrawlerImage
+from comics.aggregator.crawler import CrawlerBase
 from comics.core.comic_data import ComicDataBase
 
 
@@ -7,15 +7,10 @@ class ComicData(ComicDataBase):
     language = 'en'
     url = 'http://picturesforsadchildren.com/'
     start_date = '2007-01-01'
+    end_date = '2012-11-26'
     rights = 'John Campbell'
 
 
 class Crawler(CrawlerBase):
-    history_capable_days = 90
-    time_zone = 'US/Central'
-
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://picturesforsadchildren.com/rss')
-        for entry in feed.for_date(pub_date):
-            url = entry.summary.src('img')
-            return CrawlerImage(url)
+        pass  # Comic no longer published